# api-documentation.yaml
openapi: 3.0.0
info:
  title: Party Picker API
  version: 1.0.0
...
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /users:
    get:
      summary: Get users
      tags:
        - Users
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
          description: The email of the user
      responses:
        "200":
          description: The users were successfully retrieved
          content:
            application/vnd.users.v1+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          description: Invalid request.
    post:
      summary: Create a user
      tags:
        - Users
      requestBody:
        description: The details of the new user
        required: true
        content:
          application/vnd.user.v1+json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: The user was successfully created
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "409":
          description: The email is already in use.
  /users/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the user
    get:
      summary: Get a user by ID
      tags:
        - Users
      responses:
        "200":
          description: The user was successfully retrieved
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: The user was not found
    put:
      summary: Update a user
      tags:
        - Users
      requestBody:
        description: The new user details
        required: true
        content:
          application/vnd.user.v1+json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: The user was successfully updated
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The user was not found
    patch:
      summary: Update a user
      tags:
        - Users
      requestBody:
        description: The new user details
        required: true
        content:
          application/vnd.userVerify.v1+json:
            schema:
              type: object
              properties:
                verified:
                  type: boolean
                  description: Whether the user is verified
          application/vnd.userResetPassword.v1+json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: A null value is required to reset the password
          application/vnd.userPassword.v1+json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  minLength: 8
                  description: The new password of the user
      responses:
        "200":
          description: The user was successfully updated
          content:
            application/vnd.user.v1+json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The user was not found
  /users/{id}/occupiedDates:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the user
    get:
      summary: Get the occupied dates of a user
      tags:
        - Users
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date-time
          description: The minimum date of the occupied dates to retrieve
        - in: query
          name: to
          schema:
            type: string
            format: date-time
          description: The maximum date of the occupied dates to retrieve
      responses:
        "200":
          description: The occupied dates were successfully retrieved
          content:
            application/vnd.userOccupiedDates.v1+json:
              schema:
                type: object
                properties:
                  occupiedDates:
                    type: array
                    items:
                      type: string
                      format: date-time
        "404":
          description: The user was not found
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: The ID of the user
        name:
          type: string
          minLength: 4
          maxLength: 40
          description: The name of the user
        password:
          type: string
          minLength: 8
          format: password
          writeOnly: true
          description: The password of the user
        email:
          type: string
          format: email
          description: The email of the user
        provider:
          type: boolean
          description: Whether the user is a service provider
        averageRating:
          type: number
          readOnly: true
          description: The average rating of the user
        totalLikes:
          type: integer
          readOnly: true
          description: The total number of likes of the user
        totalEventsWorkedOn:
          type: integer
          readOnly: true
          description: The total number of events that the user worked on
        language:
          type: string
          description: The language the user prefers
        description:
          type: string
          description: The description of the user
        profilePicture:
          type: string
          format: uri
          description: The URI to the profile picture of the user
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to the user
        createdEvents:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the events of the user
        createdOfferings:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the offerings of the user
        providerRelations:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the provider relations of the user
        occupiedDates:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the occupied dates of the user
...
    FormErrors:
      type: array
      items:
        type: object
        properties:
          fieldName:
            type: string
            description: The name of the field that caused the error
          message:
            type: string
            description: The error message
...
servers:
  - description: Party Picker API
    url: http://pawserver.it.itba.edu.ar/paw-2023b-04/api